Added audio settings, fixed bug
Added a button which changes the volume by 10%. When it goes over 100%, its circles around to 0%. Added error message when the audio mixer fails to initialise (such as on Replit), and removed volume settings in that situation. Fixed bug in which holding down left click counted as multiple button presses.